+++ src/main/kotlin/com/github/anbuiii/intellijplugin/setting/AppSettingState.kt
@@ -1,0 +1,36 @@
+package com.github.anbuiii.intellijplugin.setting
+
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.State
+import com.intellij.openapi.components.Storage
+import com.intellij.util.xmlb.XmlSerializerUtil
+
+
+/**
+ * Supports storing the application settings in a persistent way.
+ * The [State] and [Storage] annotations define the name of the data and the file name where
+ * these persistent application settings are stored.
+ */
+@State(
+    name = "com.github.anbuiii.intellijplugin.setting.AppSettingsState",
+    storages = [Storage("SdkSettingsPlugin.xml")]
+)
+internal class AppSettingsState : PersistentStateComponent<AppSettingsState> {
+    var geminiKey: String = ""
+    var enable: Boolean = false
+
+    override fun getState(): AppSettingsState {
+        return this
+    }
+
+    override fun loadState(state: AppSettingsState) {
+        XmlSerializerUtil.copyBean(state, this)
+    }
+
+    companion object {
+        val instance: AppSettingsState
+            get() = ApplicationManager.getApplication().getService(AppSettingsState::class.java)
+    }
+}
+