+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Add ApiController backed by Ktor
+- Coroutine for asynchronous
+- GeminiAskQuestion action in editor popup
+- Progress Indicator for background process
+- Setting tool page